Work on satellites, add an external library, and a console. Basically, there is a file for the satellites. Can almost launch satellites, just need to add the launching part. Easy. Not.
Console only has one commnand, 'research' command. It can be shortened to 'resh', and you type the ID of the tech to research. Any tech. The game unpauses when space is pressed for the textbox, but when we have other buttons selected, it does not pause or unpause. Why can't java cooperate?

External library -- JCommons lang. Used it for the isalpha and stuff that we never ended up using.
Updated README.md.

+++ src/ConquerSpace/game/universe/GalaticLocation.java
@@ -39,7 +39,7 @@ public class GalaticLocation {
      * Get the distance.
      * @return Distance.
      */
-    public float getDistance() {
+    public int getDistance() {
         return (distance);
     }
